+++ decidim-admin/app/controllers/decidim/admin/officializations_controller.rb
@@ -14,10 +14,9 @@ module Decidim
         enforce_permission_to :read, :officialization
         @query = params[:q]
         @state = params[:state]
-
         @users = Decidim::Admin::UserFilter.for(current_organization.users.not_deleted, @query, @state)
                                            .page(params[:page])
-                                           .per(15)
+                                           .per(per_page)
       end
 
       def new